// src/generated/platform/itr.ts
export interface ItrJson {
  ITR: Itr | undefined;
}

export interface Itr {
  ITR1: Itr1 | undefined;
}

export interface Itr1 {
  CreationInfo: CreationInfo | undefined;
  FormITR1: FormItr1 | undefined;
  PersonalInfo: PersonalInfo | undefined;
  FilingStatus: FilingStatus | undefined;
  ITR1IncomeDeductions: Itr1IncomeDeductions | undefined;
  ITR1TaxComputation: Itr1TaxComputation | undefined;
  TaxPaid: TaxPaid | undefined;
  Refund: Refund | undefined;
  Verification: Verification | undefined;
}
// ...
export const ItrJson: MessageFns<ItrJson> = {
  encode(message: ItrJson, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.ITR !== undefined) {
      Itr.encode(message.ITR, writer.uint32(10).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ItrJson {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseItrJson();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.ITR = Itr.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ItrJson {
    return { ITR: isSet(object.ITR) ? Itr.fromJSON(object.ITR) : undefined };
  },

  toJSON(message: ItrJson): unknown {
    const obj: any = {};
    if (message.ITR !== undefined) {
      obj.ITR = Itr.toJSON(message.ITR);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<ItrJson>, I>>(base?: I): ItrJson {
    return ItrJson.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<ItrJson>, I>>(object: I): ItrJson {
    const message = createBaseItrJson();
    message.ITR = (object.ITR !== undefined && object.ITR !== null) ? Itr.fromPartial(object.ITR) : undefined;
    return message;
  },
};

function createBaseItr(): Itr {
  return { ITR1: undefined };
}

export const Itr: MessageFns<Itr> = {
  encode(message: Itr, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.ITR1 !== undefined) {
      Itr1.encode(message.ITR1, writer.uint32(10).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Itr {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseItr();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.ITR1 = Itr1.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Itr {
    return { ITR1: isSet(object.ITR1) ? Itr1.fromJSON(object.ITR1) : undefined };
  },

  toJSON(message: Itr): unknown {
    const obj: any = {};
    if (message.ITR1 !== undefined) {
      obj.ITR1 = Itr1.toJSON(message.ITR1);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<Itr>, I>>(base?: I): Itr {
    return Itr.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Itr>, I>>(object: I): Itr {
    const message = createBaseItr();
    message.ITR1 = (object.ITR1 !== undefined && object.ITR1 !== null) ? Itr1.fromPartial(object.ITR1) : undefined;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}